+++ src/frontend/web/src/components/Main/Modal/ModalInfo.tsx
